Fixed issue with users voting on posts without being signed in

+++ app/controllers/votes_controller.rb
@@ -24,37 +24,42 @@ class VotesController < ApplicationController
       user_votes = Vote.by_post(post).by_user(user)
     end
 
-    up = (params[:up] == 'true')
+    if user
+      up = (params[:up] == 'true')
 
-    # check if user has already voted for this post
-    if user_votes.empty?
-      if type == :blog
-        post.votes << Vote.new( :user => user, :post => post, :post_type => "BlogPost", :positive => up )
-      elsif type == :discussion
-        post.votes << Vote.new( :user => user, :post => post, :post_type => "DiscussionPost", :positive => up )
-      elsif type == :comment
-        post.votes << Vote.new( :user => user, :post => post, :post_type => "Comment", :positive => up )
-      end
-    else
-      vote = user_votes.first
-      if (up && vote.positive) || (!up && !vote.positive)
-        # user already upvoted, and clicked up again
-        # so we destroy the vote
-        # and vice versa
-        vote.destroy
+      # check if user has already voted for this post
+      if user_votes.empty?
+        if type == :blog
+          post.votes << Vote.new( :user => user, :post => post, :post_type => "BlogPost", :positive => up )
+        elsif type == :discussion
+          post.votes << Vote.new( :user => user, :post => post, :post_type => "DiscussionPost", :positive => up )
+        elsif type == :comment
+          post.votes << Vote.new( :user => user, :post => post, :post_type => "Comment", :positive => up )
+        end
       else
-        # we had an upvote, and user clicked Down
-        # so we change the upvote to a downvote
-        # or vice versa
-        vote.positive = !vote.positive
-        vote.save
+        vote = user_votes.first
+        if (up && vote.positive) || (!up && !vote.positive)
+          # user already upvoted, and clicked up again
+          # so we destroy the vote
+          # and vice versa
+          vote.destroy
+        else
+          # we had an upvote, and user clicked Down
+          # so we change the upvote to a downvote
+          # or vice versa
+          vote.positive = !vote.positive
+          vote.save
+        end
       end
-    end
 
-    @arrow_element = "#{up ? 'up-' : 'down-'}vote-#{post.id}"
-    @arrow_element_other = "#{up ? 'down-' : 'up-'}vote-#{post.id}"
-    @vote_element  = "total-votes-#{post.id}"
-    @net_votes = post.net_votes
+      @arrow_element = "#{up ? 'up-' : 'down-'}vote-#{post.id}"
+      @arrow_element_other = "#{up ? 'down-' : 'up-'}vote-#{post.id}"
+      @vote_element  = "total-votes-#{post.id}"
+      @net_votes = post.net_votes
+    else
+      @error_element = "vote-error-#{post.id}"
+      @error = "NO!"
+    end
 
     respond_to do |format|
       format.js { render :layout => false }
